EVPlug and EVState variables were not set during JSON parsing
- corrected JSON parser
- changed EVState and EVPlug from double to int

+++ EvccJSON.h
@@ -50,9 +50,9 @@ public:
 
     // nachfolgende Werte kamen vom SmartChargerXML (EDLD)
     double  getEVEvaluationPoints(void);
-    double  getEVState(void);
+    int     getEVState(void);
     double  getEVMaxPhases(void);
-    double  getEVPlug(void);
+    int     getEVPlug(void);
     double  getEVSystemEnabled(void);
     int     getEVOutput(void);
     double  getEVActualPower(void);
